Renamed GOModel to GOOrganModel
