+++ src/grandorgue/model/GOReferencePipe.cpp
@@ -13,8 +13,8 @@
 #include "config/GOConfigReader.h"
 
 #include "GOManual.h"
-#include "GOModel.h"
 #include "GOOrganController.h"
+#include "GOOrganModel.h"
 #include "GORank.h"
 #include "GOStop.h"
