+++ src/grandorgue/model/pipe-config/GOPipeConfig.cpp
@@ -9,7 +9,7 @@
 
 #include "config/GOConfigReader.h"
 #include "config/GOConfigWriter.h"
-#include "model/GOModel.h"
+#include "model/GOOrganModel.h"
 
 GOPipeConfig::GOPipeConfig(GOModel *organModel, GOPipeUpdateCallback *callback)
   : m_OrganModel(organModel),
